Authorization in some use cases

+++ src/src/Component/Application/Authorization/UseCase/DeleteUserToIssuePermissionUseCase.php
@@ -6,18 +6,26 @@ namespace LetEmTalk\Component\Application\Authorization\UseCase;
 
 use LetEmTalk\Component\Application\Authorization\Request\DeleteUserToIssuePermissionRequest;
 use LetEmTalk\Component\Domain\Authorization\Repository\UserToIssuePermissionRepository;
+use LetEmTalk\Component\Domain\Authorization\Service\AdminAuthorization;
 
 class DeleteUserToIssuePermissionUseCase
 {
     private UserToIssuePermissionRepository $userToIssuePermissionRepository;
+    private AdminAuthorization $adminAuthorization;
 
-    public function __construct(UserToIssuePermissionRepository $userToIssuePermissionRepository)
-    {
+    public function __construct(
+        UserToIssuePermissionRepository $userToIssuePermissionRepository,
+        AdminAuthorization $adminAuthorization
+    ) {
         $this->userToIssuePermissionRepository = $userToIssuePermissionRepository;
+        $this->adminAuthorization = $adminAuthorization;
     }
 
     public function execute(DeleteUserToIssuePermissionRequest $request)
     {
+        if (!$this->adminAuthorization->isAdmin($request->getUserIdentified())) {
+            throw new \InvalidArgumentException();
+        }
         $this->userToIssuePermissionRepository->delete($request->getUserId(), $request->getIssueId());
     }
